shrinker: name change
Vlad has pointed out in the past that "release_memory" makes more sense than
request_memory for the function responsible for freeing memory. He is right,
since that function will be called when we want some memory to be released. The
current names assumes that the OS is requesting memory - which is a bit
strange. It made a bit of sense when the balloon was heavily integrated into
the shrinker, but not anymore.

Signed-off-by: Glauber Costa <glommer@cloudius-systems.com>