random: Add random_harvest to the device attaching procedure
FreeBSD feeds entropy to the CPRNG in the procedure that attaches
a driver to a device. That's achieved by measuring measuring cpu
cycles during the attach operation through the function get_cyclecount.

The result from the measurement is passed over to random_harvest as the
data and only 4 bits of entropy is estimated.

Signed-off-by: Raphael S. Carvalho <raphaelsc@cloudius-systems.com>